requirements.txt
aniso8601==8.1.1
appnope==0.1.2
astroid==2.4.2
backcall==0.2.0
click==7.1.2
decorator==4.4.2
dnspython==2.1.0
email-validator==1.1.2
fhir.resources==6.0.0
Flask==1.1.2
Flask-RESTful==0.3.8
idna==3.1
ipykernel==5.4.3
ipython==7.20.0
ipython-genutils==0.2.0
isort==5.7.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
jupyter-client==6.1.11
jupyter-core==4.7.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.16
ptyprocess==0.7.0
pydantic==1.7.3
Pygments==2.7.4
pylint==2.6.0
python-dateutil==2.8.1
pytz==2021.1
pyzmq==22.0.3
six==1.15.0
toml==0.10.2
tornado==6.1
traitlets==5.0.5
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
